# Change the memory queue implementation to not pre-allocate capacity objects.