--- a/exporter/exporterqueue/bounded_memory_queue.go
+++ b/exporter/exporterqueue/bounded_memory_queue.go
@@ -16,7 +16,7 @@ import (
 // the producer are dropped.
 type boundedMemoryQueue[T any] struct {
 	component.StartFunc
-	*sizedChannel[memQueueEl[T]]
+	*sizedQueue[T]
 }
 
 // memoryQueueSettings defines internal parameters for boundedMemoryQueue creation.
@@ -29,40 +29,15 @@ type memoryQueueSettings[T any] struct {
 // callback for dropped items (e.g. useful to emit metrics).
 func newBoundedMemoryQueue[T any](set memoryQueueSettings[T]) Queue[T] {
 	return &boundedMemoryQueue[T]{
-		sizedChannel: newSizedChannel[memQueueEl[T]](set.capacity, memQueueElSizer[T]{sizer: set.sizer}),
+		sizedQueue: newSizedQueue[T](set.capacity, set.sizer),
 	}
 }
 
-// Offer is used by the producer to submit new item to the queue. Calling this method on a stopped queue will panic.
-func (q *boundedMemoryQueue[T]) Offer(ctx context.Context, req T) error {
-	return q.sizedChannel.push(memQueueEl[T]{ctx: ctx, req: req})
-}
-
 func (q *boundedMemoryQueue[T]) Read(_ context.Context) (uint64, context.Context, T, bool) {
-	item, ok := q.sizedChannel.pop()
-	return 0, item.ctx, item.req, ok
+	ctx, req, ok := q.sizedQueue.pop()
+	return 0, ctx, req, ok
 }
 
 // OnProcessingFinished should be called to remove the item of the given index from the queue once processing is finished.
 // For in memory queue, this function is noop.
-func (q *boundedMemoryQueue[T]) OnProcessingFinished(uint64, error) {
-}
-
-// Shutdown closes the queue channel to initiate draining of the queue.
-func (q *boundedMemoryQueue[T]) Shutdown(context.Context) error {
-	q.sizedChannel.shutdown()
-	return nil
-}
-
-type memQueueEl[T any] struct {
-	req T
-	ctx context.Context
-}
-
-type memQueueElSizer[T any] struct {
-	sizer sizer[T]
-}
-
-func (mqes memQueueElSizer[T]) Sizeof(el memQueueEl[T]) int64 {
-	return mqes.sizer.Sizeof(el.req)
-}
+func (q *boundedMemoryQueue[T]) OnProcessingFinished(uint64, error) {}