… exporter/exporterqueue/sized_channel.go
… exporter/exporterqueue/sized_channel_test.go